requirements.txt
fastapi
uvicorn
starlette-prometheus-middleware